livekit/requirement.txt
livekit-agents[deepgram]~=1.2
python-dotenv